# src/Coral.Api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Coral",
    "version": "v1"
  },
  "paths": {
...
    "/api/Library/scan/progress/{requestId}": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "GetScanProgress",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobProgress"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobProgress"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Library/search": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "Search",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultPaginatedCustomData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultPaginatedCustomData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultPaginatedCustomData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/tracks/{trackId}/logPlayback": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "LogPlayback",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Library/tracks/{trackId}/original": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "FileFromLibrary",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "head": {
        "tags": [
          "Library"
        ],
        "operationId": "FileFromLibrary",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Library/tracks/{trackId}/transcode": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "TranscodeTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "bitrate",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StreamDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/tracks/{trackId}/stream": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "StreamTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StreamDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/tracks/{trackId}/recommendations": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "RecommendationsForTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleTrackDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleTrackDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleTrackDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/albums/{albumId}/recommendations": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "RecommendationsForAlbum",
        "parameters": [
          {
            "name": "albumId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumRecommendationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumRecommendationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumRecommendationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Library/tracks/{trackId}/favorite": {
      "post": {
        "tags": [
          "Library"
        ],
        "operationId": "FavoriteTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Library"
        ],
        "operationId": "RemoveFavoriteTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Library/tracks/favorites/{playlistTrackId}/reorder": {
      "put": {
        "tags": [
          "Library"
        ],
        "operationId": "ReorderFavoriteTrack",
        "parameters": [
          {
            "name": "playlistTrackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Library/artists/{artistId}/favorite": {
      "post": {
        "tags": [
          "Library"
        ],
        "operationId": "FavoriteArtist",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Library"
        ],
        "operationId": "RemoveFavoriteArtist",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Library/albums/{albumId}/favorite": {
      "post": {
        "tags": [
          "Library"
        ],
        "operationId": "FavoriteAlbum",
        "parameters": [
          {
            "name": "albumId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Library"
        ],
        "operationId": "RemoveFavoriteAlbum",
        "parameters": [
          {
            "name": "albumId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Library/albums/paginated": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "PaginatedAlbums",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleAlbumDtoPaginatedQuery"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleAlbumDtoPaginatedQuery"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleAlbumDtoPaginatedQuery"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/artists/paginated": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "PaginatedArtists",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleArtistDtoPaginatedQuery"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleArtistDtoPaginatedQuery"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleArtistDtoPaginatedQuery"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/albums/{albumId}": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "Album",
        "parameters": [
          {
            "name": "albumId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/albums/{albumId}/tracks": {
      "get": {
        "tags": [
          "Library"
        ],
        "operationId": "AlbumTracks",
        "parameters": [
          {
            "name": "albumId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleTrackDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleTrackDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleTrackDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ScanJobProgress": {
        "required": [
          "embeddingsCompleted",
          "expectedTracks",
          "isFailed",
          "libraryId",
          "libraryName",
          "requestId",
          "startedAt",
          "tracksAdded",
          "tracksDeleted",
          "tracksUpdated"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "libraryId": {
            "type": "string",
            "format": "uuid"
          },
          "libraryName": {
            "type": "string"
          },
          "expectedTracks": {
            "type": "integer",
            "format": "int32"
          },
          "tracksAdded": {
            "type": "integer",
            "format": "int32"
          },
          "tracksUpdated": {
            "type": "integer",
            "format": "int32"
          },
          "tracksDeleted": {
            "type": "integer",
            "format": "int32"
          },
          "embeddingsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isFailed": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StreamDto": {
        "required": [
          "link"
        ],
        "type": "object",
        "properties": {
          "link": {
            "type": "string"
          },
          "transcodeInfo": {
            "$ref": "#/components/schemas/TranscodeInfoDto"
          },
          "artworkUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}